Improve parsing error messages.
Some were just not that easy to read but, mostly I put the AstLocator to
work, finding better positions to better showcase where the error
occurs.

In the Parser this doesn't help this much yet, since we are just about
to create AST nodes so we cannot use the Locator that much, but in
a couple of places it already really improves readability.