Add build-in print function (for strings only)
This commit required a lot of internal work, like adding function calls
to the typechecker and some refactoring in the scope.

Also I renamed Object to MooseObject because Object wasn't that explicit
and is now much clearer that this datatype belongs to moose and not to
swift.

I also fixed the missing Float type in MooseType (We declared it in the
documentation but it was nowhere to be found in the code).