Improve typechecker error highlighting
We now really put the AstLocator to work and can highlight most type
errors perfectly.
However, since type declarations for variables (or keywords like mut)
are not part of the AST we cannot highlight them. Maybe in the future
they can be part of the AST or we will live with slightly worse error
highlights.