Implement first interpreter (only assignment)
So this one had a lot of setup work. It is also kinda useless since
there are no function calls and so you cannot read the values
you set.

However the follwing example already works:
```
a = 3
b = a
```

For development purposes we currently print the environment after each
line in the REPL.

Very happy with the progress so far, build in functions are next.