Add out of order function execution to the interpreter
Before this we could only call functions that are written before the
call. Now we first explore the global scope, discover all functions
and add them to the global environment before executing anything.

This is very similary to the typechecker where we need to do
something similar to validate programs.