Add a bunch of builtin calculations and comparisons for float, int and bool.
That was a emotional roller coaster. On the one hand we now have
somewhat useful operations on integers and floats, but on the other hand
I discovered so many shortcomings of our language, which we didn't see
when we designed it. Also there are quite a lot of features that are
straight out impossible with our current design.