Improve missing function and variable error messages.
The callable error message now also shows the types required, which
makes more sense because there might be a `print` function, just not
for the types you called it with.

In general I think the typechecker could be even more helpful here, but
for now it is already quite usable.