Make every expression referable
Uff, that was a difficult one. Make them all referable revealed some
pretty intense bug, which took now almost three days to solve.

But I am kinda happy with the solution of closing environments.
I think the better solution is to make members and methods
getable but we will need some solution for function overloading.
It is now impossible since swift can do it but it is far from trivial.