Deprecate ASTLocator
The locator was a good enough solution for code highlighting but we are
aiming for a supreme experience and so it is now time to let it go.

The main reason is that there is just not enough information in the AST
to do perfect highlighting just from the AST. So we now add a location
to each AST node. With that the parser can (in cases where it is needed)
set a specific location. Therefore we can have always character-perfect
error highlighting.

This is a work in progress, the ASTLocator is already deprecated but the
highlighting are still glitchy in many cases.